% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsbarplot.R
\name{jsbarplot}
\alias{jsbarplot}
\title{plotjs Bar Plots}
\usage{
jsbarplot(
  heights,
  names.arg = NULL,
  col = NULL,
  main = NULL,
  ylab = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL,
  aria.label = NULL,
  ...
)
}
\arguments{
\item{heights}{a vector of values describing the bars that make up the plot.}

\item{names.arg}{a vector of names to be plotted below each bar.}

\item{col}{a character string with the color for bars. This can be either a hex value or the name of an R built-in color.}

\item{main}{a main title for the plot.}

\item{ylab}{a label for the y axis.}

\item{width}{width of the widget to create for the plot. The default is NULL, which results in automatic resizing based on the plot's container.}

\item{height}{height of the widget to create for the plot. The default is NULL, which results in automatic resizing based on the plot's container.}

\item{elementId}{Use an explicit element ID for the widget, rather than an automatically generated one.}

\item{aria.label}{a character string set as the \code{aria-label} attribute on the chart's canvas element for accessibility.}

\item{...}{arguments passed from methods.}
}
\description{
Creates a bar plot using 'Chart.js'.
}
