% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jspie.R
\name{jspie}
\alias{jspie}
\title{plotjs Pie Charts}
\usage{
jspie(
  x,
  labels = names(x),
  col = NULL,
  slice.text = "pct",
  donut = FALSE,
  legend.position = "right",
  main = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{x}{a vector of non-negative numerical quantities. The values in x are displayed as the areas of pie slices.}

\item{labels}{character vector giving names for the slices.}

\item{col}{character vector of colors to be used in filling the slices. Can be a hex value or an R built-in color name.}

\item{slice.text}{\code{"pct"} to display percentage-formatted values inside pie slices or \code{"id"} to display the slice's name from \code{labels}.}

\item{donut}{logical; should the chart be rendered as a donut chart instead of a full pie?}

\item{legend.position}{Position of the legend. Possible values are \code{"right"}, \code{"bottom"}, \code{"inset"}, and \code{"hide"}.}

\item{main}{a main title for the plot.}

\item{width}{width of the widget to create for the plot. The default is NULL, which results in automatic resizing based on the plot's container.}

\item{height}{height of the widget to create for the plot. The default is NULL, which results in automatic resizing based on the plot's container.}

\item{elementId}{Use an explicit element ID for the widget, rather than an automatically generated one.}
}
\description{
Draw an interactive pie chart using 'Chart.js'
}
\examples{

pie.sales <- c(0.12, 0.3, 0.26, 0.16, 0.04, 0.12)
names(pie.sales) <- c("Blueberry", "Cherry",
                    "Apple", "Boston Cream", "Other", "Vanilla Cream")
jspie(pie.sales)
jspie(pie.sales, col = c("purple", "violetred1", "green3",
                     "cornsilk", "cyan", "white"))
jspie(pie.sales, donut = TRUE)

}
