% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotjsScatter.R
\name{plotjs.default}
\alias{plotjs.default}
\title{The Default plotjs Scatterplot Function}
\usage{
\method{plotjs}{default}(
  x,
  y,
  type = "p",
  xlim = NULL,
  ylim = NULL,
  main = NULL,
  xlab = NULL,
  ylab = NULL,
  zoom = TRUE,
  col.group = NULL,
  col = NULL,
  legend.title = NULL,
  legend.position = "right",
  sci.x = FALSE,
  sci.y = FALSE,
  ...
)
}
\arguments{
\item{x}{the x coordinates for the plot.}

\item{y}{the y coordinates for the plot.}

\item{type}{1-character string giving the type of plot desired. The following values are possible: "p" for points, "l" for lines, and "b" for both points and lines.}

\item{xlim}{the x limits (x1, x2) of the plot.}

\item{ylim}{the y limits of the plot.}

\item{main}{a main title for the plot.}

\item{xlab}{a label for the x axis, defaults to a description of x.}

\item{ylab}{a label for the y axis, defaults to a description of y.}

\item{zoom}{logical; should the zooming feature be enabled for the plot?}

\item{col.group}{optionally, a factor the same length as \code{x} by which to group and color points.}

\item{col}{The colors for the lines and points. If \code{col.group} is specified, this can be a vector of colors to use for each group in the data. If \code{NULL}, Chart.js default colors are used.}

\item{legend.title}{a title for the legend. Defaults to a description of \code{col.group} if not set. You can also use \code{legend.title = FALSE} to suppress the legend title.}

\item{legend.position}{Position of the plot legend. Possible values are "right", "bottom", "inset", and "hide". This is ignored if all points are colored the same.}

\item{sci.x}{logical indicating whether scientific notation should be used for the x-axis.}

\item{sci.y}{logical indicating whether scientific notation should be used for the y-axis.}

\item{...}{arguments passed to \code{\link[htmlwidgets:createWidget]{htmlwidgets::createWidget()}}: \code{width}, \code{height}, and \code{elementId}. These arguments default to NULL. You may also pass \code{aria.label}, a character string set as the \code{aria-label} attribute on the chart's canvas element for accessibility.}
}
\description{
Draw an interactive scatterplot or line plot using 'Chart.js'.
}
\examples{
data(mtcars)
plotjs(mtcars$hp, mtcars$qsec)
plotjs(mtcars$disp, mtcars$hp, main = "Displacement vs. HP in mtcars")
}
