% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsbarplot.R
\name{plotjs.factor}
\alias{plotjs.factor}
\title{Factor Variable Bar Plots}
\usage{
\method{plotjs}{factor}(x, ylab = "Count", ...)
}
\arguments{
\item{x}{a factor.}

\item{ylab}{a label for the y axis.}

\item{...}{arguments passed to \code{\link{jsbarplot}}.}
}
\description{
This function implements a bar plot method for \code{factor} arguments to the \code{\link{plotjs}} generic function. Bar heights will be counts for the factor levels.
}
\examples{
mtcars <- mtcars
mtcars$cyl <- as.factor(mtcars$cyl)
plotjs(mtcars$cyl)
}
\seealso{
\code{\link[plotjs:jsbarplot]{jsbarplot()}}
}
